Use pathlib again; require at least PyPy3 5.7
This reverts commit 1722828ec518a31e70dfed63ebbe2defc801bc67. The
pathlib package is part of the standard library since Python 3.4; for
versions before that it can be installed from PyPI.

But PyPy3 5.5, which is compatible with Python 3.3, can't use pathlib
because it requires scandir — itself part of the standard library since
Python 3.5 — which fails to install under PyPy3 because it uses the C
API. By requiring PyPy3 5.7, which is compatible with Python 3.5, we can
ensure that both pathlib and scandir are available to PyPy3.

To support this on Travis, tox needs to run in their Trusty build
environment (see https://github.com/travis-ci/travis-ci/issues/7507#issuecomment-307529980)

All these versions make my head hurt.